fix: remediate remaining clang-tidy violations (batch 7)
- Lower quality-gate baseline from 1783 to 384 (78.5% reduction)
- Fix hicpp-use-equals-delete: move deleted members to public in 3 headers
- Fix hicpp-signed-bitwise: use unsigned shift counts in types.h
- Fix misc-const-correctness: add const to ~50 variables across source files
- Fix misc-include-cleaner: remove unused includes, add NOLINT suppression
  for platform::Mutex/sleep_for from thread.h abstraction layer
- Fix bugprone-implicit-widening-of-multiplication-result in net_impl.h
- Fix performance-inefficient-vector-operation: add reserve() in 3 files
- Fix hicpp-move-const-arg: remove std::move from trivially-copyable types
- Fix modernize-use-auto and modernize-use-emplace in sd_message/tcp_transport
- Add clang-tidy to Allure reporting via JUnit XML conversion script

+++ include/platform/posix/net_impl.h
@@ -159,7 +159,7 @@ static inline int someip_set_socket_timeout(someip_socket_t fd, int optname,
                                             int timeout_ms) {
     struct timeval tv{};
     tv.tv_sec  = timeout_ms / 1000;
-    tv.tv_usec = (timeout_ms % 1000) * 1000;
+    tv.tv_usec = static_cast<long>(timeout_ms % 1000) * 1000L;
     return ::setsockopt(fd, SOL_SOCKET, optname, &tv, sizeof(tv));
 }
